Reformatted packages for managing with setup.py
Setup.py : Thanks to Tom H.Miller from googlecl project
the setup.py is working and is easy to working with.

Moved all the helper files away from the script (fu.py)
The helperfiles went into the lib package in site the fu folder

The test files too.

There is a little md5 check to ensure that we are copyin the right
files over.

The setup will install a command called fu in your local bin dir
and save a record of the installed files to build/files.txt

this might have to be changed in the future.

Makefile : updated to smooth the install and uninstall process

sudo make install - installs and records the install files
sudo make uninstall - unistall will pipe record to rm and remove files